gbin: report row count mismatch (zStd-jni bug)
There is sometimes a bug in GaiaTools that results in failing to
read all rows from a GBIN file.

Rarely the iterator in gaia.cu1.tools.dal.gbin.GbinReader,
hence in uk.ac.starlink.gbin.GbinObjectReader, can fail to
iterate over all the rows in a GBIN file.
The correct number of rows is reported by the getTotalElementCount()
method in gaia.cu1.tools.dal.gbin.GbinMetaData, and hence in
uk.ac.starlink.gbin.GbinMeta.
This is apparently a bug in version 1.3.4-4 of the zStd-jni library
used with GaiaTools and can only be corrected by supplying a
fixed zStd-jni library.  Even when the broken library is present,
the problem only shows up for a small proportion of GBIN files.
This information comes from discussion with Alexander Hutton
(GaiaTools author, CU1).

Prior to this commit, the GbinStarTable read by GbinTableBuilder,
if GbinTableProfile.isReadMeta() returns true, would report
getRowCount() as one value, but the RowSequence it supplied
might stop after fewer rows than that.
Now, if that happens, the RowSequence throws an IOException with
a suitable message in the case that fewer than the correct number
of rows are returned.