pds: cope with zero-length fields
Fix DelimitedPds4StarTable so that fields with empty string content
don't result in a "NumberFormatException: empty String".
Scalar fields now just yield a null value, which makes sense;
array elements are left untouched, which typically means they
might be zero, which is not ideal; but if all elements are empty
then the array result is null.