table: enforce declared row counts on output
For output formats that write the row count up front, it is Bad
(will result in invalid output files) if the actual number of rows
written does not match this declared count.
That shouldn't normally happen, but it can do if for some reason an
input table lies about how many rows it contains (getRowCount() does
not match the number of rows dispensed by getRowSequence()).

Tackle this by providing a wrapper row sequence implementation
CountCheckRowSequence which will dispense the declared number of rows
even if the underlying sequence does not.  In case of a discrepancy,
informative WARNINGs are written through the logging system.
Then make sure this checked row sequence is used in output handlers
that write their row counts up front: FITS and VOTable.