ttools: recognise "smoc" as alias for "moc" in xtype
AreaDomain.getProbableMapper will now treat "smoc" the same as
"moc" as an xtype value when trying to guess how to interpret
an area specification.  There are discussions for future versions
of DALI what is the best/correct way to specify this, but if
either of those is present for an area specification it's pretty
clear what's intended, so treat it accordingly.