table: fix f.p. truncation bug in TextTableWriters
There was a nasty bug in AbstractTextTableWriter,
from which the text, ascii and IPAC output handlers inherit,
that could truncate floating point values over a certain length,
leading for instance to output for "-1.13325405E-4F" like
"-1.13325405E-" (ASCII) or "-1.332540..." (IPAC).
I don't have positive evidence that it applied to doubles as well
as floats, but probably it would do.  This behaviour has been in place
since at least 2006 - how could it have gone unnoticed for so long?

The reason was that I was assuming that (e.g.) -Float.MIN_VALUE gave
a maximal-length representation in the same way that Integer.MIN_VALUE
does; but in fact it doesn't because the precision goes away
(it's -1.4E-45), and in any case it's not reliable that any given
constant would have non-zero decimal digits at the end of the mantissa.

Fix it by using hard-coded values for the float and double maximum
string length, got by grubbing around on the internet, specifically
https://stackoverflow.com/questions/1701055/
what-is-the-maximum-length-in-chars-needed-to-represent-any-double-value