table: modify TextTableWriter col width determination
Change the way that AbstractTextTableWriter, and hence its subclasses
used for ASCII, Text and IPAC output, determine how many characters
across are required for output columns.

Writing tables in this way is now a two-pass process.
In the first pass, rows are read and cells formatted to find the
widest value in each column; by default all rows are read,
but there's a config option rowsSampled to reduce the number
if you trust that the early rows are representative of the later ones.
The second pass is a simple read->write loop.

The previous implementation was one-pass; the first few rows were
read, assessed for column widths, and cached; then they were written,
followed by a read->write loop over the remaining rows in the sequence.

Hence the new arrangement is less sophisticated and potentially less
efficient than the old one.  However it means that by default you
won't get hit by nasty surprises like truncated floating point value
renderings, which is something that's cropped up a few times.
And it's not very likely that people will be writing enormous
tables to these ASCII-alike tables with human-oriented formatting.
It would be possible to combine the best features of the old and
new ways of doing it, but unless I get evidence to the contrary,
I don't think it's worth the additional complexity.