datanode: remove FTP capability, commons.net dependency
Remove FtpDirectoryDataNode and make related changes in treeview/datanode
packages.  Apache commons-net is no longer required.
Treeview can no longer retrieve data over FTP.

This change was triggered by a security alert in Apache commons.net,
which supplies the classes that were being used for this; see
https://github.com/Starlink/starjava/security/dependabot/119 and
https://issues.apache.org/jira/browse/NET-711
I doubt if it's at all likely there would have been a problem,
but since (a) FTP is rarely used now and (b) I couldn't be bothered
to understand the details of the security alert, it seems best
to remove this dependency altogether.

The old, nominally vulnerable, version of commons.net is still
present as part of ant, but that's only used at build time,
and not likely to be talking FTP, so I don't consider that a risk.
I don't want to mess with those, very old, jar files if I can avoid it.