ttools: use Surround not Insets to model plot borders
Use the new class Surround to describe the space required round the
outside of a plot instead of java.awt.Insets.
Surround adds information about what's happening in the corners,
allowing better packing of annotations around the outside of a plot,
without wasting screen space.

At this commit it doesn't make much (any?) improvement, but it will
be more or less required to deal with "secondary" X and Y axes,
appearing on the opposite (right, top) sides to the existing
X and Y axes (left, bottom).