ttools: limit width of external legend in plots
If the legend was external to the plot and became too wide
(because too much was used to label one of the entries)
it would squeeze the width available for the plot itself
and ended up throwing an IllegalArgumentException during
surface creation.

To avoid this, limit the width of an external legend arbitrarily
to half the width of the plot.  That already looks pretty ugly,
in practice you're not going to want an external legend as big
as that.