ttools: major plot tickmarks now inside axes only
Use TickLook.STANDARD instead of TickLook.CLASSIC for drawing
ticks on axes; now major axis tickmarks distinguish themselves
from minor tickmarks by extending twice the distance inside
the plot bounds, instead of extending both inside and outside
the bounds.  This looks better if (unannotated) shadow axes are
in place, and it's more like how other plotting packages do it.

I'm not sure however that this is an improvement for Aux axes,
and maybe not for cube plot axes, so I may need to adjust those.