ttools: remove RenderingHint.VALUE_STROKE_PURE
The GraphicExporter implementation has always set the RenderingHint
KEY_STROKE_CONTROL set to VALUE_STROKE_PURE, with the intention
that graphics output to graphics files (PNGs etc), unlike to the
screen, is prioritised for correctness rather than for speed.
However, it seems that this has had always had the effect,
unnoticed by me until now, that the tickmarks on the left hand
vertical axis of plane plots are one pixel shorter than they
should be (1,3 instead of 2,4 for minor,major).
This probably reflects some geometrical fencepost error in the
axis drawing which should perhaps be fixed as well, but whatever,
it looks wrong and moreover means that the exported graphic files
do not match files plotted to the screen, which is also generally
undesirable.

So, remove this setting.  The main effect therefore is to fix the
tick mark length to look the same on all axes, as it should.
It also makes some pixel-level changes to certain other plot outputs,
but they are not obviously worse than the previous results,
for instance sky grid lines are not plotted in exactly the same
locations.

Note this only affects output to exported graphics files,
there is no change to screen display.