ttools: rationalise grid colour specification API
Modify PlaneSurface and TimeSurface so that the grid plot flag
and grid colour are combined into a single parameter when
specifying and creating a plot surface, rather than being two.
This removes an unnecessary parameter from some method signatures,
as well as making some planned upcoming changes a bit cleaner.

This does represent a (semi-)public API change, which is ceteris
paribus undesirable, but there are other breaking changes to the same
method signatures near this commit, so the additional impact is small.