ttools: add transparency to grid colour
The Plane, Time and Sky plots now provide a transparency option for their
grid colours, and Plane/Time now plots the grid over the plotted
layers rather than under it, which is what the sky plot did anyway.

I think this is an improvement.  The plane and sky plots are now
more consistent, and really it's better to have grid lines always
visible but not obscuring the plot.

However there are a couple of problems when transparency is in use
(the default).
First, the overlap of grid lines where they cross is painted twice,
hence darker than the lines themselves, which isn't really right,
though you have to be looking pretty closely to notice.
That would be quite hard to fix I think.

Secondly, PostScript output does not support transparency, so if
anybody is still exporting to PS, they will see obtrusive grid lines
in the Plane/Time plot that they didn't used to, even if they have
configured them otherwise in the UI.  That could be a change for
the worse, but I'm guessing that most people have abandoned PS for
PDF (or maybe SVG) these days.

Also, the default colour (grey) sky grid lines are less visible
than before when plotted over dark layers.