pds: use upstream fix for base16/8/2 read issue
A workaround was applied at 0759700a28 to avoid attempting to parse
ASCII_NUMERIC_BASE16, ASCII_NUMERIC_BASE8 and ASCII_NUMERIC_BASE2
types as numeric, since that caused a read failure because of a bug
in the NASA pds4-jparser library.  That workaround just treated those
fields as string values.

A fix has now been implemented in the pds4-jparser library
(https://github.com/NASA-PDS/pds4-jparser/commit/b0ad2712d96114),
currently available as v2.5.0-SNAPSHOT.  Copy the relevant source
files from that commit to this package.  Some of the other upstream
files copied here have also been changed in the NASA package, but
those changes are only cosmetic and the files here have not been
updated accordingly.  Also revert the uk.ac.starlink.pds4.FieldReader
class to its state before the workaround, so that it again treats
these fields as numeric not strings.
The RadixTextFieldAdapter class, which did something like the
upstream fix but was never used, is also removed at this commit.

Following this commit the code here now more closely matches
the latest state of the pds4-jparser library, and the behaviour
is changed so that fields of the relevant type look numeric
rather than strings when read by this library.