votable: write RESOURCE per table in fits-plus output
The VOTable metadata (content of the Primary HDU) for multi-table
FITS-plus files is now written with each TABLE element in its own
RESOURCE element, rather than writing all the TABLEs as siblings
within a single RESOURCE.

This now matches the behaviour of VOTableWriter, as altered at
ce5b49e766e8 (5 years ago), and is required for the same reason,
to ensure that DataLink-style Service Descriptors can be associated
reliably with their owner tables.  It should have been changed
at the same time but by oversight was not.

As at that commit, this introduces a format change, in this case
to the VOTable header information in FITS-plus files.
That might in principle cause problems to interoperating software,
but it probably won't (FITS-plus is in any case not really
documented as an externally usable format).

For single-table FITS-plus documents, the output is not changed.