ecsv: write empty string not "nan" for NaN output
The "nan" token, though appearing in Astropy ECSV output, is
not mentioned in the ECSV 1.0 standard, so avoid using it here.
An empty string, indicating a missing value, should do the job
just as well.

Note that the ECSV standard doesn't completely specify how to
write numeric values; the implicit assumption is that they are
done like python does it.  For most things that's the same as
Java (and other sensible languages) but for instance "inf"/"-inf"
can still appear in python output.