ecsv: adjust numeric value parsing
Make a few modifications to the way that integer and floating point
values are parsed when reading ECSV.

No longer treat the "nan" token, which may be emitted by the Astropy
ECSV writer (and earlier versions of this one) but is not mentioned
in the ECSV standard, specially.
Previously this was read as a Float.NaN/Double.NaN.
Now the behaviour is just to return null for all numeric fields
that cannot be interpreted as numeric values.

The implementation has also changed so that a fast check is
performed on all numeric fields to see if they look numeric
before attempting to parse them.  That avoids (expensive)
NumberFormatExceptions for non-numeric tokens.
As well as "nan", that behaves the same for "null", which is not
documented anywhere, but was used as a non-standard missing
value marker in the Gaia DR3 bulk download ECSV files.
The new implementation therefore runs substantially faster
than the previous one when reading Gaia DR3 ECSV files
(typically twice as fast).  There is a small (5%) performance
hit when reading files that have no non-numeric fields.

Also replace e.g. Long::parseLong with Long::valueOf, which avoids
an implicit unboxing operation (probably no effect on behaviour,
but it's more respectable).