topcat: report DL1.1 columns in Datalink view
Modify the Row Detail panel in the Datalink table view to
display a couple of additional columns introduced in DataLink 1.1:
content_qualifier and local_semantics.

To avoid using up more screen space, rework the display a bit so that
the additional information is displayed on existing lines along with
existing related items.