ant: fix platform class list to include jdk.**
Fix the hard-coded list of package prefixes to include "jdk" in the
ant runtime JRE.
This corresponds to a fix that was made in ant 1.9.8, described at
https://bz.apache.org/bugzilla/show_bug.cgi?id=59556 and
https://github.com/apache/ant/pull/19/files

Without this fix, attempting to build the TTOOLS and TOPCAT packages
with Java 9 produced a java.lang.NoClassDefFoundError for class
jdk.internal.reflect.ConstructorAccessorImpl when attempting
reflective construction during the autodocumentation build phase
(uk.ac.starlink.util.ObjectFactory.createObject).
(An alternative workaround was to set fork="yes" in the relevant
ant targets).

The inclusion of the modified bootstrap ant.jar file in this
commit is not necessary, since it has to be rebuilt before building
the rest of the starjava packages, but since it includes the
offending class I've included it in this commit for consistency
and to preserve the sanity of anybody who goes digging into these
files in the future.

Obviously, updating ant to a later version should also do the trick,
but I am reluctant to make that change in case it breaks something.