ttools: some tidying of Doclet classes
The superclass XmlDoclet is removed since it's never used alone,
and all the functionality is put into FullXmlDoclet.
A couple of other bits of tidying done at the same time.

These classes are scheduled for retirement in any case, but these
changes make it clearer how the subsequent refactoring relates
to the earlier implementation.