ttools: refactor doclets used for JEL documentation
Rework the doclet implementations used for producing user-facing
documentation of the classes in uk.ac.starlink.ttools.func and
uk.ac.starlink.topcat.func that are available from the JEL
expression language.

The implementation is now split into one class that parses the
javadocs (SunDoclet) and a DocletOutput interface that takes
care of formatting the parsed information into some useful output
(with implementations for SUN-friendly XML and
MethodBrowser-friendly HTML).  The formatting is therefore now
completely isolated from the javadoc parsing.

That is necessary for compatibility with later JDKs, because the javadoc
parsing is currently done using the com.sun.javadoc API, which is
OK at Java 8, but deprecated at Java 9 and withdrawn completely
at Java 17.  In order to build these javadocs at future JDKs
it should just be necessary to swap out SunDoclet for a class
that works with the later jdk.javadoc.doclet API.

The XML and HTML output at this commit has been checked to be
byte-for-byte exactly the same as before, for the currently
existing files on which they operate (uk.ac.starlink.ttools.func
and uk.ac.starlink.topcat.func), under JDK8 and JDK9.