ttools: add JdkDoclet class
A new doclet class JdkDoclet, using the jdk.javadoc.doclet API, is added.

This class is a drop-in replacement for the existing SunDoclet,
which uses the com.sun.javadoc API, and produces almost identical
output, at least for those classes currently being documented
in this way (uk.ac.starlink.ttools.func.*, uk.ac.starlink.topcat.func.*).
The only discrepancies are whitespace differences in HTML links.

To use it, fix the build file so that this class is not excluded
from compilation and references to the class SunDoclet are
replaced with references to JdkDoclet.

At Java 8 the new class JdkDoclet will not compile.
At Java 9-16, both SunDoclet and JdkDoclet will compile, run, and
produce the same output, though the libraries used by SunDoclet are
deprecated.  At Java 17 and later, SunDoclet will not compile.

At this commit, the build file explicitly excludes this file
from compilation so that compiling with JDK 8 will not fail.