ttools: ensure doclet and runtime use same mangle syntax
Fix it so that the javadoc documentation document names used at
runtime by MethodBrowser, as acquired from DocNames, are
guaranteed to use the same syntax as the doclet that produces
these names.  Previously the two used the same conventions,
but this was not forced or documented.

The type names used in javadoc resource names are also
simplified, they now use only the inner class name not the
outer.inner classname for internal classes.  There is also
one place (field types) where the generated output has changed
in the same way.  This only affects inner classes
(currently only in KCorrections).

This makes the filenames a bit less unwieldy in those cases,
but the real reason is because it seems to be hard to get the
full classnames (with $ in the right places) for inner classes
when using the jdk.javadoc API.