ttools: fix build file to work with doclet variants
The TTOOLS build file will now only attempt to build the JdkDoclet
if it is running in a Java9+ JDK, since the required library classes
are not present at 1.8 or earlier.  So it should all just work at
any JDK version earlier than Java 17, at which SunDoclet will not
compile.  It would be possible to add more logic to prevent attempting
to build SunDoclet at Java 17+.

Additionally there is a property jel.doclet that can be set in both
TTOOLS and TOPCAT build files that will let you switch to using the
JdkDoclet rather than the SunDoclet if you want.  This will work for
Java9+, but at present is set to use SunDoclet, which works for any
JDK up to 16.