ttools: fix scaling in MultiPointForms
Previously, the extent of MultiPointForm shapes such as vectors
and ellipses was calculated on the basis only of unscaled input
values, and graphical scaling related to input units etc was
applied after the data coordinates had been calculated.
That's OK for linear figures like a Cartesian vector,
but wrong for sky forms; the input extents were being interpreted
as degrees and spherical geometry performed accordingly, which
works very badly if for instance your input values are in mas/yr
and have values of >360.

This serious bug has been present for ever in the sky vector/ellipse
plotting code, and gone unnoticed so far because the testing I've done
was for relatively small angles (<<360 in whatever units I was using).

This commit reworks the MultiPointForm to do the manipulations
correctly; quite a bit of reorganisation was required.
The abstract MultiPointForm class now has two concrete subclasses,
one for Cartesian and one for Sky geometry, and these provide
separate CoordSet interfaces to accomodate the needs of the
different geometries.
This allows the interpretation of input angular extents to be
separated from scaling of figures in graphics coordinates,
which is necessary for proper handling of autoscaling on the sky.
Behaviour should be pretty similar to previously for small
input angular extents on the sky, but it will fix erroneous
behaviour for large extents.