util: handle HTTP 308s in the same way as 307s
The HTTP 308 response code was introduced in RFC 7538 (2015) and
is described alongside other 3xx redirects in RFC 9110.
Although RFC 7538 says that services should not rely on clients
understanding 308 because of its relatively recent introduction,
it is known to be in use at least by SDSS skyserver.

There are some loose ends in redirection handling that this commit
fails to tidy up.  That is because the relevant code is scheduled
for replacement in future by the much delayed authentication rework
which ought to do things properly.