ttools: fix multi-marker plot threading bug
Fix a bug in DataPosSequence.split which did the splitting wrong
(silly mistake) in the case of multi-SubCloud PointClouds.
The effect was that when iterating over graphics positions for
marker-like shapes with multiple positions (like Mark2) some of
the positions for each row were omitted when in multithreaded mode.
In practice that meant for instance that Mark2 only plotted one
not two markers for large tables (small ones don't usually get split).

This is a regression bug since the introduction of multithreaded
plotting.  Bit surprised it hasn't been noticed before now.