auth: modify context success annotation in AuthManager
Rather than using two lists contexts_ and reserveContexts_ to store
contexts which are known to be OK and maybe-OK respectively,
instead maintain one list of TestedContext objects which have
additional mutable metadata describing their success history.

Both before and after this change the logic here is hairy.
I'm not confident that the logic is exactly what's wanted still,
but I think it allows for better behaviour than the way it did
it before.