ttools: fix LineTracer bug
Fix a bug in LineTracer that was flushing the vertex buffer every
time two vertices were on top of each other.  That meant especially
that dashing didn't work at all well for lines with some points
that were close together on the plotting surface, but also probably
that line joins were not so good and it may also have affected
performance.

This was a regression bug introduced when the colour argument was
added to the addVertex method at commit f20ffe73b4.