vo: rework/extend Service tab LanguageFeature display
The display of the TAPRegExt capabilities document in the
TAP window Service sub-tab now has two additional entries,
"ADQL 2.1 Optional Features" and "Non-Standard Language Features".

The code that presents these, along with the UDF and Geometry
features, has been generalised and simplified.
It no longer explicitly caters for non-ADQL languages in the
TAPRegExt, but nobody seems to have any of those anyway.