tjoin: fix PairMode.BEST so it's really symmetric
One of the steps in the "Best match, symmetric" pair match mode
was not actually symmetric in the two input tables.
The initial accumulation of pair links done by RowMatcher.findAllPairs
included only the best match in table R(andom) for each row of
table S(equential) rather than all the rows.
That meant that result of the subsequent
RowMatcher.eliminateMultipleRowEntries step, and hence the
output of the match, was dependent on which input table was R
and which was S.
This dependency wasn't obvious from using the applications,
since the code assigned the input tables to R and S based on their
content, so presenting the tables in a different order to
stilts/topcat wouldn't have made a difference to the output.

Following this commit the processing is symmetric and a bit
more stable and predictable than its previous behaviour,
though in crowded regions it's not really obvious that the
new answers are better than the old ones (though they probably are).

This problem looks like it was introduced a long time ago,
at commit 8a0d735f9c in 2010 (STILTS v2.1-2/TOPCAT v3.5-2).
It was noticed now (by Georgios Vasilopoulos in Strasbourg)
because of a regression issue: STILTS v3.4-6 gave a different
result on a particular find=best pair match than STILTS v3.4-7.
The reason for that was the improvement in (sky) coverage handling
that made the code smarter at filtering out sources which couldn't
possibly match, which in turn happened to change the order in which
it decided to assign the input tables as R and S (since the one with
fewer possible matches was different after this change).

The reason the problem hasn't been noticed before now, and only
showed up as a regression issue, is because of the obscure and
enigmatic semantics of the BEST PairMode - it's not clear by
looking at the results whether it's done the right thing or not.
BEST is really best avoided in crowded regions (and it behaves
the same as ALL outside of crowded regions) and it should maybe
be deprecated.  Unfortunately it has always been the default
option in TOPCAT.  I keep considering doing something about that.