tjoin: ensure TupleDecoders supply legal sky positions
The SkyCoverage.TupleDecoders supplied to SkyMatchEngine
CoverageFactory instances need to supply physically reasonable
positions (latitude must be in the range -PI/2<=delta<=+PI/2)
otherwise the coverage assessment done by cds.healpix.Healpix
can fail with an IllegalArgumentException and scupper the match.
Add additional checks to make sure that this is the case.

This fixes a regression issue introduced when Coverage was
introduced at d15b547a93.