topcat: post subset edit window on bad edited expression
Move the logic for checking and updating the expression for a
synthetic RowSubset out of SubsetWindow and into SubsetQueryWindow.
If the expression update fails (e.g. gnu.jel.CompilationException)
the window is made visible so that the expression can be edited.

This makes for a better user experience because if a little mistake
is made in the new expression, the text is still available
(in the QueryWindow Expression field) rather than just getting wiped
out so you have to start again.