vo: ADQL2.1 standard to PR 20230418
This is mostly a documentation thing; not much change is required,
since most of the ADQL version-specific work is done by the VOLLT
library and not the starjava code.  But it's useful to make clear
that we are in principle dealing with a recent revision of ADQL
(hopefully quite close to the eventual REC).

This is slightly provisional since the document PR-ADQL-2.1-20230418
is expected to go on the document repository but is not there yet;
other late changes might modify the date and possibly content,
but not likely in a way that affects the code here.
Once that document is in place at www.ivoa.net, the URL in
vo/src/etc/adql_hints.xhtml should be updated.