ttools: fix huge table rendering bug in SwingMode
The CellRenderers were getting reset when using SwingMode with a
ViewHugeTableModel (the setModel caused the JTable to reset them).
The effect was that e.g. NaNs showed up as a funny character and
decimal points weren't aligned.  Put them back by hand.

This doesn't seem to happen with the corresponding usage of
ViewHugeTableModel in TOPCAT, presumably because the TableColumns
and/or renderers are managed differently there.
So no corresponding change in topcat code, but maybe keep an eye
out for it if modifying this behaviour in future.