vo: update hard-coded registry URLs http->https
The RegTAP and RI1.0 registry URLs for the Euro-VO and STSci registries
were using HTTP, but also exist at HTTPS equivalents.  Update the
hard-coded values to use the HTTPS equivalents.  These are what show
up in e.g. the topcat registry search dialogue URL selectors.

The STSci one actually works for either at present.
The Euro-VO one doesn't work any more at HTTP, since the HTTP
response is a 302 redirect, which doesn't work for POSTs,
which is what mostly gets done for registry access.
So this change is necessary for the Euro-VO one to work.