vo: no DAL submit trigger on SkyPositionEntry action
The SkyPositionEntry had the SkyDalTableLoadDialog Submit action as
an ActionListener.  That meant that if you (a) hit return,
or (b) a sky position was resolved in any of its fields (RA, Dec, Radius),
it would try to initiate the DAL query and, often, give you an
annoying popup window because you hadn't filled in the Radius
field yet.

It's not really helpful to initiate the query on either of the
conditions (a) or (b), so just stop an action here triggering Submit.
The user can/should initiate the query by hitting the OK button.

Note that trigger on (b) was only introduced relatively recently
(TOPCAT v4.8-7/commit 002e7d0c3f4), so this looks a bit like a
regression fix - annoying popup when hitting the Resolve button
only happened since then.