ttools: fix LayerSpec datasys/viewsys issues
When topcat uses LayerSpecs to generate a STILTS command for
display and export, it is tricky to get the assignment of
viewsys and datasys right.  The viewsys is the sky system in
which the plot is viewed, but other sky systems may be chosen
by plot layers, and whether they are or are not explicitly
listed needs some special handling.

Till now it worked OK for data plots, but not for SkyGrid and
Healpix plots, which also have datasys-like config options.
Generalise the datasys-like config option handling so it does
the right thing.

This bugfix was stimulated by a report from Robin Geyer.