vo: fix qualified identifier bug in VOLLT invocation
The recent upgrade to VOLLT pre-2.0 introduced a bug in ADQL parsing
caused by errors in the way that table metadata is presented to
the VOLLT parser.  I had misunderstood the semantics of the (new)
adql.db.DBTable.isCaseSensitive method, which really indicates
whether the table's name is a delimited identifier.
The effect was that references to table names using delimited
identifiers were incorrectly flagged as parse errors.  Now fixed.