ttools: move Messenger vocabulary definition
Creation of the MESSENGER instance of VocabChecker is moved from
EpnTapStage to VocabChecker to keep these things in one place and
make it easier to check vocabulary content (currently done using
the main method of VocabChecker).  Took the opportunity to update
the vocabulary at the same time, it now contains a term "Radio"
and it's not Preliminary any more.