ttools: generalise and fix multizone annotation
Replace the xAnnotate and yAnnotate flags in PlaneAxisAnnotation
with a new structure SideFlags which encodes whether annotation
is required on all four axes, not just two.

This is required for correct axis annotation in the presence of
secondary axes.  Before this change, secondary X axes in stacked
time plots didn't work properly - primary and secondary X annotations
were suppressed on all plots except the bottom one, when what
should happen is that the primary X annotations should be
suppressed on all but the bottom, while secondary X annotations
should be suppressed on all but the top.

This generalisation will also provide the capabilities to work with
multizone plots having configurations other than a vertical stack.