ttools: LayerType getPositionCount -> getCoordGroup
Replace the getPositionCount method of LayerType with getCoordGroup,
which can be used to get the same information, but has wider use.
This is required by some later commits.  It's slightly dodgy since
there is not necessarily a well-defined CoordGroup for each LayerType,
but in practice something representative can be returned.