vo: manage VOLLT ivoid case-sensitivity
Fix the handling of TAPRegExt language features read from the TAP
capabilities file so that the registry part of the feature types
("ivo://ivoa.net/std/TAPRegExt") can be given with any capitalisation.
If it differs from the above (canonical?) form, VOLLT doesn't
recognise the feature type.  That is arguably a VOLLT misfeature,
but in any case work around it here by modifying the feature types
so that VOLLT sees the case-sensitive string it is looking for
during LanguageFeature construction, regardless of how it was
specified in the VOSI capabilities.

Feature types are written in all-lower-case in the current ADQL 2.1 PR,
and have been observed in some TAP/ADQL 2.1 services, which is how
I came to notice this issue.  Prior to this fix, language features
were not being recognised by the ADQL validation, so that queries
using declared optional ADQL features were erroneously flagged invalid.