vo: use new Ivoid class to model IVO IDs
Represent VO Identifiers in library code using the Ivoid class
rather than as Strings, which handles the case-sensitivity of
equality correctly.  In nearly all cases just lowercasing the
whole thing would do the trick, but we might as well do it properly.

The requisite case-folding was in any case not done everywhere it
should have been, and typing the standard constants in this way has
makes it easier to identify where that's happening and put it right.

In particular taplint was incorrectly handling case-sensitivity
in IVOIDs (e.g. language features) and generating spurious reports.
This should now be fixed.  There may be other similar issues fixed too.

This is rather a large commit and there are opportunities for getting
parts of it wrong; for instance Collection.contains(String) on a
Collection<Ivoid> won't do the right thing, but is not picked up
by the compiler.  So it's possible that this commit has introduced
bugs related to identifier matching at the same time as fixing
other ones.  Look out.