ttools: use withScroll consistently in PlotPlacement
The createPlotDecorations method of PlotPlacement did not have a
withScroll parameter, and just behaved as if that value was false.
That worked with the pre-Surround/secondary axis code, but it is
required in general in the current implementation.
Add this parameter and fill it in as required in calling code.