ttools: fix sky axis Surround calculations
The plot Surround was not calculated correctly for SkySurface,
meaning that the Aux axis and Legend were positioned in the
wrong place and jumped around during scrolling.
This was done wrong when the Surround class was introduced to
support secondary axes, so the observed incorrect behaviour dates
from then (one release ago).  Now calculated more carefully and
original behaviour restored.