ttools: refactor plot2 zone definitions
Change the way that the information specifying what to plot in each
zone of a (single- or multi-zone) plot is passed around within the
plot2 classes.

The class ZoneContent which defines what's in a plot zone still
contains a PlotLayer[] array, but it now also contains the
Profile and Aspect that together with a SurfaceFactory define
the details of the PlotSurface.  The members defining legend and
title placement have been moved out into a new class Trimming.

The other changes in this commit follow from the above;
where a PlotLayer[], Profile and Aspect are passed around together
methods now use a ZoneContent, and where a LegendIcon, legend position
and title are passed around together, methods now use a Trimming.

This makes some of the code more transparent and tidier, but the
real purpose is to decouple the decorations defined in Trimmings
(legend, title and potentially other things in future) from the
zone content (surface and layer definition).  So far, there is
always a one-to-one mapping between ZoneContent and Trimming,
but this commit will make it easier to break that assumption if
more general multi-zone plotting arrangements are required in future.