topcat: remove incorrect assertion in SquarePusher
It was asserted that during a mouse drag operation there must be
a recorded drag start point, as long as the component was enabled.
This isn't true, the start point will be null also if the drag
started outside of the target.

So remove the assertion.  This won't make much difference to anything.