topcat: fix broken matched subset assignment
There was a nasty bug in topcat's pair match window, which assigns a
new subset "matched" to matched rows in a match table when it's useful
to do so (some but not all rows result from a successful match).
Although the rest of the match worked fine, the contents of this
subset was completely wrong (it had the right number of entries,
but they were the wrong ones).

This is a regression bug since topcat v4.8-4 (1471c96266),
which implemented matching performance improvements.
The problem was that I was looking at the unsorted row links
rather than the sorted ones that go to make the output table.
Now fixed.

This bug was reported by Claire Greenwell (Durham).