ttools: fix multi-cloud plot threading bug (again)
When multiple point clouds were handled by DataPosSequence
in multi-threading mode, some clouds could get ignored.
This bug manifested for instance by causing the topcat re-autoscaling
to select a region based on only one subset rather than on multiple
subsets when multiple subsets were plotted.  Also in failing to plot
some markers in multiple-marker layer types like Mark2.

This bug was tackled at 1211af26b but the fix that time was incorrect,
it just fiddled about in a way that masked it for some tests rather
than doing what was required, which was to ensure that the new cloud
was initialised properly in all cases.  This commit therefore correctly
(I hope) fixes a regression bug introduced with multithreaded plotting.